Add giganto-client crate
- Replace quic communication logic with giganto-client api
- Replace the raw data structure with the structure of giganto

Closes: #226
