+++ Cargo.toml
@@ -11,8 +11,8 @@ chrono = "0.4"
 clap = { version = "4.1", features = ["wrap_help"] }
 csv = "1.1"
 ctrlc = { version = "3", features = ["termination"] }
+giganto-client = { git = "https://github.com/aicers/giganto.git", rev = "c1d610c"}
 lazy_static = "1.4"
-num_enum = "0.5"
 num-traits = "0.2"
 quinn = "0.9"
 regex = "1"
