--- src/zeek/smtp.rs
@@ -1,104 +1,0 @@
-use super::{parse_zeek_timestamp, TryFromZeekRecord};
-use anyhow::{anyhow, Context, Result};
-use serde::Serialize;
-use std::net::IpAddr;
-
-#[derive(Debug, Serialize)]
-pub(crate) struct ZeekSmtp {
-    src_addr: IpAddr,
-    src_port: u16,
-    dst_addr: IpAddr,
-    dst_port: u16,
-    proto: u8,
-    duration: i64,
-    mailfrom: String,
-    date: String,
-    from: String,
-    to: String,
-    subject: String,
-    agent: String,
-}
-
-impl TryFromZeekRecord for ZeekSmtp {
-    fn try_from_zeek_record(rec: &csv::StringRecord) -> Result<(Self, i64)> {
-        let time = if let Some(timestamp) = rec.get(0) {
-            parse_zeek_timestamp(timestamp)?.timestamp_nanos()
-        } else {
-            return Err(anyhow!("missing timestamp"));
-        };
-        let src_addr = if let Some(src_addr) = rec.get(2) {
-            src_addr
-                .parse::<IpAddr>()
-                .context("invalid source address")?
-        } else {
-            return Err(anyhow!("missing source address"));
-        };
-        let src_port = if let Some(src_port) = rec.get(3) {
-            src_port.parse::<u16>().context("invalid source port")?
-        } else {
-            return Err(anyhow!("missing source port"));
-        };
-        let dst_addr = if let Some(dst_addr) = rec.get(4) {
-            dst_addr
-                .parse::<IpAddr>()
-                .context("invalid destination address")?
-        } else {
-            return Err(anyhow!("missing destination address"));
-        };
-        let dst_port = if let Some(dst_port) = rec.get(5) {
-            dst_port
-                .parse::<u16>()
-                .context("invalid destination port")?
-        } else {
-            return Err(anyhow!("missing destination port"));
-        };
-        let mailfrom = if let Some(mailfrom) = rec.get(8) {
-            mailfrom.to_string()
-        } else {
-            return Err(anyhow!("missing mailfrom"));
-        };
-        let date = if let Some(date) = rec.get(10) {
-            date.to_string()
-        } else {
-            return Err(anyhow!("missing date"));
-        };
-        let from = if let Some(from) = rec.get(11) {
-            from.to_string()
-        } else {
-            return Err(anyhow!("missing from"));
-        };
-        let to = if let Some(to) = rec.get(12) {
-            to.to_string()
-        } else {
-            return Err(anyhow!("missing to"));
-        };
-        let subject = if let Some(subject) = rec.get(17) {
-            subject.to_string()
-        } else {
-            return Err(anyhow!("missing subject"));
-        };
-        let agent = if let Some(agent) = rec.get(23) {
-            agent.to_string()
-        } else {
-            return Err(anyhow!("missing agent"));
-        };
-
-        Ok((
-            Self {
-                src_addr,
-                src_port,
-                dst_addr,
-                dst_port,
-                proto: 0,
-                duration: 0,
-                mailfrom,
-                date,
-                from,
-                to,
-                subject,
-                agent,
-            },
-            time,
-        ))
-    }
-}
